feat(lean-imt)!: switched to passing the hash function as a function instead of an address
Switched to passing the hash function as a function instead of an address and fixed some of the
readability of _insertMany.

Also added a regular insert to the insertMany test in order to catch issues with self.sideNodes
being off.

BREAKING CHANGE: switched to passing the hash function as a function

instead of an address